[update]レスポンシブデザインの修正
個別質問ページ公開機能のレスポンシブ対応を、コンポーネント毎入れ替えるように変更
